Update the layout of the nurse page and rename the files
All the buttons are working properly except for the Search button in the search page, the Update button in the patient history and the Save button in the vitals.
