--- PatientMedicalHistory.java
@@ -1,5 +1,0 @@
-package application;
-
-public class PatientMedicalHistory {
-
-}
