--- src/application/PatientHistory.fxml
@@ -1,39 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.CheckBox?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.control.TextArea?>
-<?import javafx.scene.layout.AnchorPane?>
-<?import javafx.scene.text.Font?>
-<?import javafx.scene.text.Text?>
-<?import javafx.scene.text.TextFlow?>
-
-<AnchorPane prefHeight="542.0" prefWidth="910.0" xmlns="http://javafx.com/javafx/22" xmlns:fx="http://javafx.com/fxml/1" fx:controller="application.PatientHistoryController">
-   <children>
-      <Label layoutX="350.0" layoutY="54.0" prefHeight="32.0" prefWidth="190.0" text="Medical History">
-         <font>
-            <Font size="24.0" />
-         </font>
-      </Label>
-      <Button fx:id="backToMenuButton" layoutX="30.0" layoutY="30.0" mnemonicParsing="false" onAction="#switchToMenu" prefHeight="40.0" prefWidth="161.0" text="Back to Menu" />
-      <Button fx:id="logOutButton" layoutX="727.0" layoutY="34.0" mnemonicParsing="false" onAction="#logOut" prefHeight="40.0" prefWidth="128.0" text="Log Out" />
-      <Text fx:id="name" layoutX="61.0" layoutY="132.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Name:" wrappingWidth="153.0498046875" />
-      <Text fx:id="dob" layoutX="216.0" layoutY="133.0" strokeType="OUTSIDE" strokeWidth="0.0" text="DOB:" wrappingWidth="89.0498046875" />
-      <Text fx:id="height" layoutX="350.0" layoutY="133.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Height:" />
-      <Text fx:id="weight" layoutX="559.0" layoutY="132.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Weight:" wrappingWidth="89.0498046875" />
-      <TextFlow layoutX="182.0" layoutY="193.0" prefHeight="200.0" prefWidth="200.0" AnchorPane.topAnchor="193.0" />
-      <Text fx:id="allegies" layoutX="41.0" layoutY="194.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Allegies:" />
-      <Text fx:id="illiness" layoutX="167.0" layoutY="194.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Illiness:" wrappingWidth="69.3017578125" />
-      <Text fx:id="medications" layoutX="291.0" layoutY="194.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Medications:" wrappingWidth="118.0498046875" />
-      <Text fx:id="immunizations" layoutX="426.0" layoutY="194.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Immunizations:" wrappingWidth="265.0498046875" />
-      <CheckBox fx:id="immunizationCheckBox" layoutX="559.0" layoutY="173.0" mnemonicParsing="false" onAction="#immunizationCheck" text="up-to-date" />
-      <Text fx:id="ImunizatinMissing" layoutX="708.0" layoutY="194.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Missing:" wrappingWidth="76.0498046875" />
-      <Label layoutX="100.0" layoutY="225.0" prefHeight="28.0" prefWidth="134.0" text="Past Check-in" />
-      <TextArea fx:id="pastCheckInArea" layoutX="61.0" layoutY="264.0" prefHeight="259.0" prefWidth="200.0" />
-      <Label layoutX="382.0" layoutY="225.0" prefHeight="28.0" prefWidth="126.0" text="Family History" />
-      <TextArea fx:id="familyHistoryArea" layoutX="350.0" layoutY="268.0" prefHeight="252.0" prefWidth="200.0" />
-      <Label layoutX="692.0" layoutY="225.0" text="Surgeries" />
-      <TextArea fx:id="surgeryArea" layoutX="639.0" layoutY="264.0" prefHeight="252.0" prefWidth="200.0" />
-   </children>
-</AnchorPane>
