+++ src/application/checkInController.java
@@ -1,0 +1,5 @@
+package application;
+
+public class checkInController {
+
+}
